+++ Router/route.py
@@ -208,7 +208,7 @@ class Route:
             return -1
 
         if self.offset + direction >= len(self.route):
-            return self.offset
+            return len(self.route) - 1
 
         self.offset += direction
         return self.offset